Move configPath to config
