+++ spec/services/user/session.spec.ts
@@ -1,6 +1,6 @@
 import fs from "fs-extra";
 import { describe, expect, it } from "vitest";
-import { configPath } from "../../../src/services/config/paths.js";
+import { configPath } from "../../../src/services/config/config.js";
 import { readSession, writeSession } from "../../../src/services/user/session.js";
 
 describe("session", () => {
