+++ src/services/config/paths.ts
@@ -1,6 +1,5 @@
 import path from "node:path";
 import { fileURLToPath } from "node:url";
-import { config } from "./config.js";
 
 /**
  * Returns the filename of the current module.
@@ -45,21 +44,3 @@ export const workspacePath = (...segments: string[]): string => path.join(worksp
  * @returns The absolute path to the file or directory.
  */
 export const assetsPath = (...segments: string[]): string => workspacePath("assets", ...segments);
-
-/**
- * Returns an absolute path within the {@linkcode config.configDir}
- * directory.
- *
- * @param segments - The segments of the path to join.
- * @returns The absolute path to the file or directory.
- */
-export const configPath = (...segments: string[]): string => path.join(config.configDir, ...segments);
-
-/**
- * Returns an absolute path within the {@linkcode config.homeDir}
- * directory.
- *
- * @param segments - The segments of the path to join.
- * @returns The absolute path to the file or directory.
- */
-export const homePath = (...segments: string[]): string => path.join(config.homeDir, ...segments);
